Move syntax error guard from shared capability probes to parser utils
- relocate isSyntaxErrorWithLocation from src/shared/utils/capability-probes.js to src/parser/src/utils/syntax-error-guards.js so parser-focused error helpers live alongside the parser

- update parser entrypoint, plugin facade, and tests to consume the new module
