+++ src/parser/gml-parser.js
@@ -5,3 +5,6 @@ export {
     sanitizeConditionalAssignments,
     applySanitizedIndexAdjustments
 } from "./src/conditional-assignment-sanitizer.js";
+export {
+    isSyntaxErrorWithLocation
+} from "./src/utils/syntax-error-guards.js";
